+++ spec/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141215204615) do
+ActiveRecord::Schema.define(version: 20141216165853) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -36,7 +36,7 @@ ActiveRecord::Schema.define(version: 20141215204615) do
     t.datetime "created_at"
     t.datetime "updated_at"
     t.integer  "arm_id"
-    t.integer  "moderator_id", null: false
+    t.integer  "moderator_id"
   end
 
   create_table "memberships", force: true do |t|
@@ -65,6 +65,7 @@ ActiveRecord::Schema.define(version: 20141215204615) do
     t.datetime "last_sign_in_at"
     t.inet     "current_sign_in_ip"
     t.inet     "last_sign_in_ip"
+    t.boolean  "is_admin"
   end
 
   add_index "participants", ["email"], name: "index_participants_on_email", unique: true, using: :btree